+++ board/config.h
@@ -35,7 +35,8 @@
 #elif defined(STM32F2) || defined(STM32F4)
   #include "stm32fx/stm32fx_config.h"
 #else
-  // pass, building for tests
+  // building for tests
+  #include "fake_stm.h"
 #endif
 
 #endif
